+++ src/io/save_gals_binary.h
@@ -105,7 +105,7 @@ extern "C" {
       float Concentration;
       float mdot_cool;
       float mdot_stream;
-      float g_max;
+      double g_max;
     };
 
     /* Proto-Types */
